+++ assets/scripts/main.js
@@ -1,3 +1,24 @@
+/* General Variables & Presets
+/===================================================== */
+
+	/* Viewport Width
+	/------------------------*/
+	var $vpWidth = jQuery(window).width();
+
+	/* Touch Device
+	/------------------------*/
+	var $root = $('html');
+	var isTouch = 'ontouchstart' in document.documentElement;
+	if (isTouch) {
+		$root.attr('data-touch', 'true');
+	} else {
+		$root.attr('data-touch', 'false');
+	}
+
+
+/* Functions
+/===================================================== */
+
 /**
  * Mobile Navigation handler
  *